requirements.txt
PyMySQL==1.0.2
python-dotenv==0.20.0
pandas==1.4.3
numpy==1.23.0